+++ JBQQuizMe.ViewModel/PreferenceKeys.cs
@@ -5,5 +5,9 @@ namespace JBQQuizMe.ViewModel
     public static class PreferenceKeys
     {
         public const string ReadQuestions = "read_questions";
+
+        public const string StartQuestionRange = "start_question_range";
+
+        public const string EndQuestionRange = "end_question_range";
     }
 }